Moved markdown to helpers
